Update chapter0.md
I would like to offer a suggestion for Q6. As triplet primes must occur over a sequence of 6 numbers three of these numbers must be divisible by two and in addition at least one different member of the sequence must be divisible by 3. As three is a new prime in 3,5,7 it doesn't affect it here but you will find that once the lowest number in the sequence is above three, the odd multiple of three stops any chance of getting triplet primes.